Consolidated all write_signal functions
