+++ clarity/evaluator/msbg/msbg_utils.py
@@ -10,7 +10,6 @@ from typing import Final, TypedDict
 import numpy as np
 import scipy
 import scipy.signal
-import soundfile
 from numpy import float64, ndarray
 from soundfile import SoundFile
 
@@ -566,45 +565,3 @@ def read_signal(
         )
 
     return signal
-
-
-def write_signal(
-    filename: str | Path,
-    signal: ndarray,
-    sample_rate: float,
-    floating_point: bool = True,
-    strict: bool = False,
-) -> None:
-    """Write a signal as fixed or floating point wav file.
-
-    NB: setting 'strict' to True will raise error on overflow. This would be
-    a more natural default but it would break existing code that did not
-    check for overflow.
-
-    Args:
-        filename (str|Path): name of file in to write to.
-        signal (ndarray): signal to write.
-        sample_rate (float): sampling frequency.
-        floating_point (bool): write as floating point else an ints (default: True).
-        strict (bool): raise error if signal out of range for int16 (default: False).
-    """
-
-    if sample_rate != MSBG_FS:
-        logging.warning(
-            f"Sampling rate mismatch: {filename} with "
-            f"sample frequency = {sample_rate}."
-        )
-        # raise ValueError("Sampling rate mismatch")
-
-    if floating_point is False:
-        subtype = "PCM_16"
-        # Signal is float and we want to convert to int16
-        # *NB* Not  *= in next line as we need to make a copy
-        signal = signal * 32768
-        if strict and (np.max(signal) > 32767 or np.min(signal) < -32768):
-            raise ValueError("Signal out of range -1.0 to 1.0")
-        signal = signal.astype(np.dtype("int16"))
-    else:
-        subtype = "FLOAT"
-
-    soundfile.write(filename, signal, sample_rate, subtype=subtype)
