+++ recipes/cec1/baseline/evaluate.py
@@ -12,8 +12,9 @@ from tqdm import tqdm
 from clarity.evaluator.mbstoi.mbstoi import mbstoi
 from clarity.evaluator.mbstoi.mbstoi_utils import find_delay_impulse
 from clarity.evaluator.msbg.msbg import Ear
-from clarity.evaluator.msbg.msbg_utils import MSBG_FS, pad, read_signal, write_signal
+from clarity.evaluator.msbg.msbg_utils import MSBG_FS, pad, read_signal
 from clarity.utils.audiogram import Audiogram
+from clarity.utils.file_io import write_signal
 
 
 def listen(ear, signal, audiogram_l, audiogram_r):
