+++ tests/evaluator/msbg/test_msbg_utils.py
@@ -10,8 +10,8 @@ from clarity.evaluator.msbg.msbg_utils import (
     pad,
     read_gtf_file,
     read_signal,
-    write_signal,
 )
+from clarity.utils.file_io import write_signal
 
 # import scipy
